requirements/base.txt
# Pure Python runtime dependencies shared by Pi and development machines.

numpy>=1.26.4,<2.0.0
Pillow>=10.3.0,<12.0.0
psutil>=5.9.8,<8.0.0
pyttsx3>=2.90,<3.0.0
pytesseract>=0.3.10,<0.4.0
sounddevice>=0.4.6,<1.0.0
tenacity>=8.2.3,<10.0.0
typing-extensions>=4.11.0,<5.0.0

